Use uppercase methods in README
HTTP methods are case-sensitive. The examples in the README currently
lead developers to think that this API normalizes all of the method
names to uppercase, leading to issues like #37 and #254.

Setting the right precedent in the README will prevent developers
from making mistake of trying to use method: 'patch' (or any
other verb that isn't normalized for backwards compatibility).