Change Headers multiple value handling for spec compatibility
Consider this Headers object:

    var h = new Headers()
    h.append('accept', 'text/html')
    h.append('accept', 'text/plain')
    h.append('content-type', 'application/json')

Before:

- `h.get('accept')` returned `text/html`
- `h.getAll('accept')` returned an array of values
- `h.forEach` (and other iterables) did distinct iterations for each
  value of the same header

Now:

- `h.get('accept')` returns `text/html,text/plain`
- `h.getAll()` is no more
- `h.forEach` (and other iterables) now only do one iteration for each
  headers name, regardless of multiple values

This is in accordance with Section 3.1.2 of the spec, "combine" concept.

The updated tests currently break in Chrome and Firefox when exercising
the native implementation because their implementation is outdated. The
implementation in Edge is more correct, but the tests still fail there
because its implementation combines values with `, ` (notice the space)
rather than `,`.