Add support for DataView as POST body
DataView is one of the possible views for ArrayBuffer, which the fetch
spec says should be supported. However, IE 10-11 don't handle it well
when passed unmodified to XMLHttpRequest. As a workaround, wrap every
DataView in a Blob.

It's interesting to note that IE 10's string representation of a
DataView instance is `[object Object]` rather than `[object DataView]`.
Therefore, the only way to detect it is via `isPrototypeOf`.