Use new exports to clean up test suite preparation
Prior to this change, `test.js` had to be loaded before `fetch.js` so it
can nulify the native implementation of `window.fetch` to force testing
of the polyfill. Now that we have exports from the UMD file, we can
abandon this hack and simply access the exports to juggle between native
and polyfilled versions of the implementation.